Add sections, conform styling, comment headers
This adds the ability to define a "Projects" section, as well as a
"Certifications" section. Additionally, it generally makes the TeX code
easier to read and understand by applying some common formatting and
adding comment headers for various sections of the template. Finally, it
changes some of the styling in the template generally to improve
aesthetics.